fix(shared): cast usage-stat token SUMs to float8 to avoid int4 overflow (#166)
The dashboard /dashboard route 500'd with Postgres "integer out of
range" once lifetime usage accumulated. getUsageStats, getSessionStats,
and getDailyByModel cast their SUM(token/duration) aggregates to ::int
(int4, max 2,147,483,647). Each per-row column is int4 and in-range, but
the aggregate SUM is not bounded by int4: lifetime SUM(cache_read_tokens)
had already reached 2.24B, overflowing the cast.

Cast the token and duration SUMs to ::float8 instead (matching the
existing cost_usd treatment). float8 returns a JS number and represents
integers exactly up to 2^53, far beyond realistic token totals, so the
downstream arithmetic and types are unchanged. COUNT(*)::int stays — row
counts won't overflow.

Adds a regression test that seeds rows whose SUM crosses the int4
ceiling and asserts each aggregate path returns the correct total instead
of throwing. Reverting any one cast to ::int reproduces the original
"integer out of range".

+++ packages/shared/src/services/usage.ts
@@ -190,11 +190,11 @@ export async function getUsageStats(sinceIso?: string): Promise<UsageStats> {
   const baseSql = `SELECT
         COUNT(*)::int                          AS "turns",
         COALESCE(SUM(cost_usd), 0)::float8     AS "cost",
-        COALESCE(SUM(input_tokens), 0)::int    AS "inputRaw",
-        COALESCE(SUM(output_tokens), 0)::int   AS "output",
-        COALESCE(SUM(cache_creation_tokens),0)::int AS "cacheCreation",
-        COALESCE(SUM(cache_read_tokens), 0)::int    AS "cacheRead",
-        COALESCE(SUM(duration_ms), 0)::int     AS "duration"
+        COALESCE(SUM(input_tokens), 0)::float8 AS "inputRaw",
+        COALESCE(SUM(output_tokens), 0)::float8 AS "output",
+        COALESCE(SUM(cache_creation_tokens),0)::float8 AS "cacheCreation",
+        COALESCE(SUM(cache_read_tokens), 0)::float8 AS "cacheRead",
+        COALESCE(SUM(duration_ms), 0)::float8  AS "duration"
       FROM usage`;
   const result =
     sinceIso === undefined
@@ -255,10 +255,10 @@ export async function getDailyByModel(sinceIso?: string): Promise<DailyByModelRo
   const projection = `to_char(timestamp AT TIME ZONE current_setting('TimeZone'), 'YYYY-MM-DD') AS "day",
               COALESCE(model, 'unknown')                                 AS "model",
               COALESCE(SUM(cost_usd), 0)::float8                         AS "cost",
-              COALESCE(SUM(input_tokens), 0)::int                        AS "rawInput",
-              COALESCE(SUM(cache_creation_tokens), 0)::int               AS "cacheCreation",
-              COALESCE(SUM(cache_read_tokens), 0)::int                   AS "cacheRead",
-              COALESCE(SUM(output_tokens), 0)::int                       AS "output",
+              COALESCE(SUM(input_tokens), 0)::float8                     AS "rawInput",
+              COALESCE(SUM(cache_creation_tokens), 0)::float8            AS "cacheCreation",
+              COALESCE(SUM(cache_read_tokens), 0)::float8                AS "cacheRead",
+              COALESCE(SUM(output_tokens), 0)::float8                    AS "output",
               COUNT(*)::int                                              AS "turns"`;
   if (sinceIso === undefined) {
     const result = await pool.query<DailyByModelRow>(
@@ -338,11 +338,11 @@ export async function getSessionStats(sessionId: string): Promise<SessionStats |
     `SELECT
        COUNT(*)::int                          AS turn_count,
        COALESCE(SUM(cost_usd), 0)::float8     AS total_cost,
-       COALESCE(SUM(input_tokens), 0)::int    AS total_input,
-       COALESCE(SUM(output_tokens), 0)::int   AS total_output,
-       COALESCE(SUM(cache_creation_tokens),0)::int AS total_cache_create,
-       COALESCE(SUM(cache_read_tokens), 0)::int    AS total_cache_read,
-       COALESCE(SUM(duration_ms), 0)::int     AS total_duration,
+       COALESCE(SUM(input_tokens), 0)::float8 AS total_input,
+       COALESCE(SUM(output_tokens), 0)::float8 AS total_output,
+       COALESCE(SUM(cache_creation_tokens),0)::float8 AS total_cache_create,
+       COALESCE(SUM(cache_read_tokens), 0)::float8 AS total_cache_read,
+       COALESCE(SUM(duration_ms), 0)::float8  AS total_duration,
        MIN(timestamp)                         AS first_at,
        MAX(timestamp)                         AS last_at
      FROM usage WHERE session_id = $1`,
